+++ src/extension.ts
@@ -12,6 +12,8 @@ interface ExtensionConfig {
     keepOneEmptyLine?: boolean;
     triggerOnSave?: boolean;
     insertLineAfterBlock?: boolean;
+    allowElseAfterBlock?: boolean;
+    formatIfElseBlock?: string;
     removeBlockPadding?: boolean;
     languageIds?: string[];
 }
@@ -21,6 +23,8 @@ let config: ExtensionConfig = {
     keepOneEmptyLine: true,
     triggerOnSave: true,
     insertLineAfterBlock: true,
+    allowElseAfterBlock: true,
+    formatIfElseBlock: '1tbs',
     removeBlockPadding: true,
     languageIds: ['javascript', 'typescript', 'json']
 };
@@ -41,35 +45,120 @@ function isValidLanguage(languageId: string) {
 function processLines(lines: vscode.TextLine[]): string[] {
     return (lines || [])
         .reduce((a: any[], line) => {
-            let prevLine: vscode.TextLine = a.slice(-1)[0];
+            const lineTrimmed = line.text.trim();
+
+            let prevLine: any = a.slice(-1)[0];
             if (prevLine) {
-                if (prevLine.isEmptyOrWhitespace && line.isEmptyOrWhitespace) { return a; }
+                const prevLineText = prevLine.text !== undefined ? prevLine.text : prevLine;
+                const prevLineTrimmed = prevLineText.trim();
+                // If previous line and current line are empty, skip
+                if (prevLine.isEmptyOrWhitespace && line.isEmptyOrWhitespace)
+                    return a;
 
                 if (config.removeBlockPadding) {
-                    if (line.isEmptyOrWhitespace && prevLine.text.trim()[prevLine.text.trim().length - 1] === '{') { return a; }
-                    if (prevLine.isEmptyOrWhitespace && beginsWithClosingBracket(line.text)) { a.pop(); }
+                    // If current line is empty and previous line ended with '{', skip
+                    if (line.isEmptyOrWhitespace && prevLineTrimmed[prevLineTrimmed.length - 1] === '{')
+                        return a;
+
+                    // If previous line is empty and current line starts with '}', remove previous line
+                    if (prevLine.isEmptyOrWhitespace && startsWithClosingBrace(line.text))
+                        a.pop();
                 }
 
-                if (config.insertLineAfterBlock && prevLine.text.trim() === '}' && !line.isEmptyOrWhitespace &&
-                    !beginsWithClosingBracket(line.text) && !beginsWithClosingTag(line.text)) { a.push(null); }
+                if (config.formatIfElseBlock === '1tbs') {
+                    // If previous line starts with '}' and current line starts with 'else', concatenate
+                    if (prevLineTrimmed === '}' && startsWithElse(line.text)) {
+                        a.pop();
+                        a.push(prevLineText + ' ' + lineTrimmed);
+                        return a;
+                    }
+                }
+
+                if (config.formatIfElseBlock === '1tbs' || config.formatIfElseBlock === 'stroustrup') {
+                    // If previous line is if or else and current line is opening brace, concatenate
+                    if (isIfOrElseWithoutOpenBrace(prevLineTrimmed) && lineTrimmed === '{') {
+                        a.pop();
+                        a.push(prevLineText + ' ' + lineTrimmed);
+                        return a;
+                    }
+                }
+
+                if (config.insertLineAfterBlock) {
+                    // If previous line starts with '}' and current line is not a closing brace/tag, add a blank line
+                    if (prevLineTrimmed === '}' && !line.isEmptyOrWhitespace && !startsWithClosingBraceOrTag(line.text)) {
+                        if (!config.allowElseAfterBlock || !startsWithElse(line.text))
+                            a.push(null);
+                    }
+                }
             }
 
-            if (config.keepOneEmptyLine !== true && line.isEmptyOrWhitespace) { return a; }
+            if (config.formatIfElseBlock === 'stroustrup' || config.formatIfElseBlock === 'allman') {
+                // If line starts with '} else', split
+                if (lineTrimmed.startsWith('} else')) {
+                    const index = line.text.indexOf('}');
+                    a.push(line.text.substr(0, index + 1));
+                    a.push(line.text.substr(0, index) + line.text.substr(index + 2));
+                    return a;
+                }
+            }
+
+            if (config.formatIfElseBlock === 'allman') {
+                // If line has other text and ends with '{', split
+                if (isIfOrElseWithOpenBrace(lineTrimmed)) {
+                    a.push(line.text.substr(0, lineTrimmed.length - 1));
+                    a.push(line.text.substr(0, line.firstNonWhitespaceCharacterIndex) + '{');
+                    return a;
+                }
+            }
+
+            // If line is empty, skip
+            if (config.keepOneEmptyLine !== true && line.isEmptyOrWhitespace)
+                return a;
 
             return a.concat([line]);
-        }, []).map(line => line ? line.text : '');
+        }, []).map(line => {
+            if (!line)
+                return '';
+
+            if (line.text !== undefined)
+                return line.text;
+
+            return line;
+        });
 }
 
-function beginsWithClosingBracket(text: string) {
+function startsWithClosingBrace(text: string) {
     const trimmed = text.trim()[0];
     if (trimmed.length === 0) { return false; }
     return trimmed === '}' || trimmed === ']' || trimmed === ')';
 }
 
-function beginsWithClosingTag(text: string) {
+function startsWithClosingTag(text: string) {
     const trimmed = text.trim();
     if (trimmed.length === 0) { return false; }
-    return trimmed[0] === '<' && trimmed[1] === '/';
+    return trimmed.startsWith('</');
+}
+
+function startsWithClosingBraceOrTag(text: string) {
+    return startsWithClosingBrace(text) || startsWithClosingTag(text);
+}
+
+function startsWithElse(text: string) {
+    const trimmed = text.trim();
+    if (trimmed.length === 0) { return false; }
+    return trimmed.startsWith('else');
+}
+
+function isIfOrElseWithOpenBrace(text: string) {
+    const trimmed = text.trim();
+    return (trimmed.startsWith('if (') || trimmed.startsWith('else') || trimmed.startsWith('} else')) &&
+        trimmed.endsWith('{');
+}
+
+function isIfOrElseWithoutOpenBrace(text: string) {
+    const trimmed = text.trim();
+    return (trimmed.startsWith('if (') || trimmed.startsWith('else') || trimmed.startsWith('} else')) &&
+        (trimmed.endsWith(')') || trimmed.endsWith('else'));
 }
 
 function selectLines(editor: vscode.TextEditor, start: number, end: number) {
@@ -96,7 +185,7 @@ function doAction(event: string) {
 
     // Select start and end lines
     var selection = editor.selection;
-    var start = 1;
+    var start = 0;
     var end = editor.document.lineCount;
     if (selection.start.line !== selection.end.line) {
         start = selection.start.line;