+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+# v0.8.0
+
+* allowElseAfterBlock: If set to true won't insert a blank line after block statement if next line is an else
+* formatIfElseBlock: Defines if and how to format if-else blocks
+
 # v0.5.1
 
-* fix: insertLineAfterBlock ignores closing tags (like </script>, useful for Vue)
-
+* fix: insertLineAfterBlock ignores closing tags (like `</script>`, useful for Vue)
 
 # v0.2.0
 
@@ -16,10 +20,13 @@
 * add: react support (updated default configuration)
 
 # v0.1.0
+
 * fix: performance issues
 
 # v0.0.2
+
 * Minor bug fix.
 
 # v0.0.1
+
 * Initial version